explicit types for vec_split_i and vec_split_s created common_support and alloocator files
